+++ packages/core/src/languages/javascript/index.ts
@@ -58,7 +58,7 @@ export type JavaScriptLanguagePluginOptions = {
  * });
  *
  * @since v0.1.0
- * @version v0.1.19
+ * @version v0.1.20
  */
 export default class JavaScriptLanguagePlugin extends SimpleLanguagePlugin {
   /**