+++ packages/core/CHANGELOG.md
@@ -10,6 +10,7 @@ Versioning].
 
 - Add option to ignore strings to `SingleGroupMangleExpression`.
 - Fix mangling query selectors in nested CSS selectors.
+- Fix unintended mangling of some query selectors.
 
 ## [0.1.19] - 2021-04-30